moved BedAssignment model to dedicated module
